[ty] Remove flattened uv metadata deserialization
Deserialize workspace and dependency fields into one struct so parse errors
retain their field locations. Keep dependency conversion in a child module
without exposing the metadata fields.
