+++ crates/ty_project/src/uv/metadata.rs
@@ -1,5 +1,7 @@
+use std::collections::BTreeMap;
 use std::path::PathBuf;
 
+use compact_str::CompactString;
 use pep440_rs::Version;
 use ruff_db::system::{System, SystemPath, SystemPathBuf};
 use ruff_ranged_value::{RangedValue, ValueSource};
@@ -7,8 +9,9 @@ use serde::Deserialize;
 use thiserror::Error;
 use ty_python_semantic::dependency::DependencyMetadata;
 
-use super::dependencies::WorkspaceDependencies;
 use crate::metadata::python_version::SupportedPythonVersion;
+
+mod dependencies;
 
 #[derive(Debug, Clone, PartialEq, Eq, get_size2::GetSize)]
 pub(crate) struct UvMetadata {
@@ -16,7 +19,10 @@ pub(crate) struct UvMetadata {
     members: Box<[WorkspaceMember]>,
     environment: Option<SystemPathBuf>,
     python_version: Option<RangedValue<SupportedPythonVersion>>,
-    dependencies: WorkspaceDependencies,
+    schema: Schema,
+    workspace: Option<NodeReference>,
+    resolution: BTreeMap<CompactString, ResolutionNode>,
+    module_owners: BTreeMap<CompactString, Box<[ModuleOwner]>>,
 }
 
 impl UvMetadata {
@@ -64,13 +70,15 @@ impl UvMetadata {
             members: metadata.members,
             environment,
             python_version,
-            dependencies: metadata.dependencies,
+            schema: metadata.schema,
+            workspace: metadata.workspace,
+            resolution: metadata.resolution,
+            module_owners: metadata.module_owners,
         })
     }
 
     pub(crate) fn dependency_metadata(&self) -> Option<DependencyMetadata> {
-        self.dependencies
-            .to_metadata(&self.workspace_root)
+        self.to_dependency_metadata()
             .inspect_err(|error| {
                 tracing::debug!(
                     "Skipping dependency checks for '{}': {error:#}",
@@ -86,6 +94,7 @@ pub(crate) struct WorkspaceMember {
     pub(crate) name: Box<str>,
     /// Directory containing the member's `pyproject.toml`.
     pub(crate) path: SystemPathBuf,
+    id: CompactString,
 }
 
 #[derive(Debug, Error)]
@@ -147,14 +156,24 @@ fn resolve_python_version(
     Ok(RangedValue::new(version, ValueSource::UvMetadata))
 }
 
+/// The uv metadata used to discover the workspace and check imports against its dependencies.
+///
+/// See uv's [schema documentation] and [serialization types] for the upstream format.
+///
+/// [schema documentation]: https://docs.astral.sh/uv/reference/internals/metadata/#schema
+/// [serialization types]: https://github.com/astral-sh/uv/blob/main/crates/uv-resolver/src/lock/export/metadata.rs
 #[derive(Deserialize)]
 struct WorkspaceMetadata {
     workspace_root: PathBuf,
     #[serde(default)]
     members: Box<[WorkspaceMember]>,
     environment: Option<WorkspaceEnvironment>,
-    #[serde(flatten)]
-    dependencies: WorkspaceDependencies,
+    schema: Schema,
+    workspace: Option<NodeReference>,
+    #[serde(default)]
+    resolution: BTreeMap<CompactString, ResolutionNode>,
+    #[serde(default)]
+    module_owners: BTreeMap<CompactString, Box<[ModuleOwner]>>,
 }
 
 #[derive(Deserialize)]
@@ -166,6 +185,57 @@ struct WorkspaceEnvironment {
 #[derive(Deserialize)]
 struct WorkspacePython {
     version: Version,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, get_size2::GetSize)]
+struct Schema {
+    version: SchemaVersion,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, get_size2::GetSize)]
+#[serde(rename_all = "snake_case")]
+enum SchemaVersion {
+    Preview,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, get_size2::GetSize)]
+struct ModuleOwner {
+    package_id: CompactString,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, get_size2::GetSize)]
+struct ResolutionNode {
+    kind: NodeKind,
+    name: Option<CompactString>,
+    source: Option<Source>,
+    // uv always emits this field, even for leaves. Missing edges are incomplete metadata, not
+    // evidence that a project has no direct dependencies.
+    dependencies: Box<[NodeReference]>,
+    #[serde(default)]
+    optional_dependencies: Box<[NodeReference]>,
+    #[serde(default)]
+    dependency_groups: Box<[NodeReference]>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, get_size2::GetSize)]
+#[serde(rename_all = "snake_case")]
+enum NodeKind {
+    Package,
+    Extra(CompactString),
+    Group(CompactString),
+    Workspace,
+    Script,
+    Build,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, get_size2::GetSize)]
+struct Source {
+    editable: Option<SystemPathBuf>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, get_size2::GetSize)]
+struct NodeReference {
+    id: CompactString,
 }
 
 #[cfg(test)]
@@ -280,9 +350,14 @@ mod tests {
                 "resolution": resolution
             });
 
-            assert_matches!(
-                UvMetadata::from_metadata(&serde_json::to_vec(&metadata)?, &system),
-                Err(UvMetadataError::InvalidMetadata(_))
+            let metadata = serde_json::to_string_pretty(&metadata)?;
+            let error = match UvMetadata::from_metadata(metadata.as_bytes(), &system) {
+                Err(UvMetadataError::InvalidMetadata(error)) => error,
+                result => anyhow::bail!("expected invalid metadata, got {result:?}"),
+            };
+            assert!(
+                error.line() > 0 && error.line() < metadata.lines().count(),
+                "expected the error to point to its field, not the end of the response: {error}"
             );
         }
